camera 到 多通道预览优化

+++ app/src/main/java/com/bisa/health/cust/view/CustomECGViewBg.java
@@ -37,22 +37,22 @@ public class CustomECGViewBg extends View {
         }
 
 
-        int width = this.getWidth();
-        int height = this.getHeight();
+        float width = this.getWidth();
+        float height = this.getHeight();
         paint.setAlpha(255);
         paint.setStrokeWidth(4);
         paint.setStyle(Paint.Style.STROKE);
-        Rect rect=new Rect(0,0,width,height);
+        Rect rect=new Rect(0,0,(int)width,(int)height);
         canvas.drawRect(rect,paint);
         paint.setStrokeWidth(1);
         float ecgDrawMinUnitWidth=height/(ECGViewConfig.CHCOUNT*ECGViewConfig.DrawDefaultScale);//刻度-2 2的格子数
         float unitXMinHeight=ecgDrawMinUnitWidth/5;//计算一个格子的像素
         paint.setAlpha(100);
-        for(int y=0;y<=ECGViewConfig.DrawDefaultScale*ECGViewConfig.CHCOUNT*5;y++){
+        for(float y=0f;y<=ECGViewConfig.DrawDefaultScale*ECGViewConfig.CHCOUNT*5;y++){
             canvas.drawLine(0, y*unitXMinHeight,width,  y*unitXMinHeight, paint);
         }
         paint.setAlpha(150);
-        for(int y=0;y<=ECGViewConfig.DrawDefaultScale*ECGViewConfig.CHCOUNT*5;y++){
+        for(float y=0f;y<=ECGViewConfig.DrawDefaultScale*ECGViewConfig.CHCOUNT*5;y++){
             if(y%5==0){
                 canvas.drawLine(0, y*unitXMinHeight,width,  y*unitXMinHeight, paint);
             }
@@ -60,12 +60,12 @@ public class CustomECGViewBg extends View {
         }
         float xCount=width/unitXMinHeight;
         paint.setAlpha(50);
-        for(int x=0;x<xCount;x++){
+        for(float x=0f;x<xCount;x++){
             canvas.drawLine(x*unitXMinHeight,0,x*unitXMinHeight, height, paint);
 
         }
         paint.setAlpha(150);
-        for(int x=0;x<xCount;x++){
+        for(float x=0f;x<xCount;x++){
             if(x%5==0){
                 canvas.drawLine(x*unitXMinHeight,0,x*unitXMinHeight, height, paint);
             }
